src/api/routes/kiosks.php src/EcclesiaCRM/APIControllers/KiosksController.php
